refactor: ProductCard 을 공통 컴포넌트로 분리
- TrendingGift, ThemeProduct 모두 grid 구조를 사용
- 따라서 ProductsGrid 를 공통 컴포넌트로 분리
- 하지만 데이터 요청 방식, 로딩 UI 가 모두 달라 공통 컴포넌트로 뽑기엔 컴포넌트의 책임과 복잡도가 높아진다고 판단
- 따라서 ProductCard 만 공통 컴포넌트로 분리
