+++ src/components/ProductCard/types/productGridTypes.ts
